Create year-month-day-format
This code shows how to build a simple time format with years - months - days. Using a starting date to a end date without applying any library or 3rd party libraries. The code is written with java script only. HTML is also used in order to display the output of the logic.